bugfix - fix make dashd pathing

+++ bin/_install_dashd.sh
@@ -1,34 +1,42 @@
 #!/bin/bash
 
-echo "installing dashd"
+DOWNLOAD_TYPE="RPi2" # or linux64 or linux32
 
 if [ -e ~/.dash/dash.conf ]; then
     exit 0
 fi
 
+echo "installing dashd"
+
 mkdir -p ~/.dash/testnet/testnet3
 cp -f bin/.dash.conf.template ~/.dash
 cd ~/.dash
 touch dashd.pid testnet/testnet3/dashd.pid
-wget https://www.dash.org/binaries/dash-0.12.0.56-RPi2.tar.gz
-tar zxvf dash-0.12.0.56-RPi2.tar.gz
+if [ ! -e dash-0.12.0.56-${DOWNLOAD_TYPE}.tar.gz ]; then
+    wget https://www.dash.org/binaries/dash-0.12.0.56-${DOWNLOAD_TYPE}.tar.gz
+fi
+tar zxvf dash-0.12.0.56-${DOWNLOAD_TYPE}.tar.gz
 ln -s dash-0.12.0/bin/dash-cli .
 ln -s dash-0.12.0/bin/dashd .
 export PATH=~/.dash:$PATH
 echo 'export PATH=~/.dash:$PATH' >> ~/.bashrc
 
-wget https://raw.githubusercontent.com/UdjinM6/dash-bootstrap/master/links.md -O links.md
-MAINNET_BOOTSTRAP_FILE=$(head -1 links.md | awk '{print $11}' | sed 's/.*\(http.*\.zip\).*/\1/')
-wget $MAINNET_BOOTSTRAP_FILE
-unzip ${MAINNET_BOOTSTRAP_FILE##*/}
-rm links.md bootstrap.dat*.zip
+if [ ! -e bootstrap.dat ]; then
+    wget https://raw.githubusercontent.com/UdjinM6/dash-bootstrap/master/links.md -O links.md
+    MAINNET_BOOTSTRAP_FILE=$(head -1 links.md | awk '{print $11}' | sed 's/.*\(http.*\.zip\).*/\1/')
+    wget $MAINNET_BOOTSTRAP_FILE
+    unzip ${MAINNET_BOOTSTRAP_FILE##*/}
+    rm links.md bootstrap.dat*.zip
+fi
 
 cd testnet/testnet3
-wget https://raw.githubusercontent.com/UdjinM6/dash-bootstrap/master/linksTestnet.md -O linksTestnet.md
-TESTNET_BOOTSTRAP_FILE=$(head -1 linksTestnet.md | awk '{print $11}' | sed 's/.*\(http.*\.zip\).*/\1/')
-wget $TESTNET_BOOTSTRAP_FILE
-unzip ${TESTNET_BOOTSTRAP_FILE##*/}
-rm linksTestnet.md bootstrap.dat*.zip
+if [ ! -e bootstrap.dat ]; then
+    wget https://raw.githubusercontent.com/UdjinM6/dash-bootstrap/master/linksTestnet.md -O linksTestnet.md
+    TESTNET_BOOTSTRAP_FILE=$(head -1 linksTestnet.md | awk '{print $11}' | sed 's/.*\(http.*\.zip\).*/\1/')
+    wget $TESTNET_BOOTSTRAP_FILE
+    unzip ${TESTNET_BOOTSTRAP_FILE##*/}
+    rm linksTestnet.md bootstrap.dat*.zip
+fi
 
 # build confs
 function render_conf() {
@@ -38,8 +46,7 @@ function render_conf() {
         eval echo "$REPLY"
     done < .dash.conf.template > $1
 }
+cd ../..
 render_conf dash.conf
 render_conf testnet/dash.conf
 echo "testnet=1" >> testnet/dash.conf
-dashd
-dashd -datadir=testnet
